+++ src/base/operations/create_atomic_swap_bid_request_builder.js
@@ -1,7 +1,12 @@
 import { default as xdr } from '../generated/xdr_generated'
 import { BaseOperation } from './base_operation'
 import { UnsignedHyper } from 'js-xdr'
-import isUndefined from 'lodash/isUndefined'
+import {
+  validateAmount,
+  validateAssetCode,
+  validateCreatorDetails,
+  validateUint64
+} from '../../utils/validators'
 
 export class CreateAtomicSwapBidRequestBuilder {
   /**
@@ -18,21 +23,10 @@ export class CreateAtomicSwapBidRequestBuilder {
      * @returns {xdr.Operation}
      */
   static createAtomicSwapBidRequest (opts) {
+    this._validateCreateAtomicSwapBidOp(opts)
+
     let rawRequest = {}
-    if (!BaseOperation.isValidAmount(opts.baseAmount)) {
-      throw new Error('opts.amount is invalid')
-    }
-    rawRequest.baseAmount = BaseOperation._toUnsignedXDRAmount(
-      opts.baseAmount)
-
-    if (!BaseOperation.isValidAsset(opts.quoteAsset)) {
-      throw new Error('opts.quoteAssets is invalid')
-    }
-
-    if (isUndefined(opts.creatorDetails)) {
-      throw new Error('opts.creatorDetails is undefined')
-    }
-
+    rawRequest.baseAmount = BaseOperation._toUnsignedXDRAmount(opts.baseAmount)
     rawRequest.quoteAsset = opts.quoteAsset
     rawRequest.creatorDetails = JSON.stringify(opts.creatorDetails)
     rawRequest.askId = UnsignedHyper.fromString(opts.askID)
@@ -57,5 +51,12 @@ export class CreateAtomicSwapBidRequestBuilder {
       attrs.request().baseAmount())
     result.quoteAsset = attrs.request().quoteAsset().toString()
     result.creatorDetails = JSON.parse(attrs.request().creatorDetails().toString())
+  }
+
+  static _validateCreateAtomicSwapBidOp (opts) {
+    validateUint64({ value: opts.askID, fieldName: 'opts.askID' })
+    validateAmount({ value: opts.baseAmount, fieldName: 'opts.baseAmount' })
+    validateCreatorDetails({ value: opts.creatorDetails, fieldName: 'opts.creatorDetails' })
+    validateAssetCode({ value: opts.quoteAsset, fieldName: 'opts.quoteAsset' })
   }
 }